move utilities to util folder and create index to export modules
