+++ src/e2e/tests/utils/auth/drafterAuth.ts
@@ -1,5 +1,5 @@
 import { Page, expect } from '@playwright/test';
-import { drafterUsername, drafterPassword } from '../../env';
+import { drafterUsername, drafterPassword } from '../env/env';
 
 // Function to log in as a Drafter user
 export async function drafterLogin(page: Page) {
@@ -39,31 +39,5 @@ export async function drafterLogin(page: Page) {
     // Handle errors that occur during login
     console.error('Login failed', error);
     throw new Error('Login failed'); // Re-throw the error to fail the test
-  }
-}
-
-// Function to log out as a Drafter user
-export async function drafterLogout(page: Page) {
-  try {
-    // Click on the 'Sign Out' button
-    await page.getByRole('button', { name: 'Sign Out' }).click();
-
-    // Confirm sign out
-    await page.getByLabel('Yes, sign out').click();
-
-    // Verify that the user is redirected to the home page after logging out
-    await expect(page).toHaveURL('/');
-
-    // Navigate to the home page
-    await page.goto('/');
-
-    // Verify 'Log in with IDIR' button is displayed
-    await expect(
-      page.locator('#main-content >> role=button[name="Log in with IDIR"]'),
-    ).toHaveText('Log in with IDIR');
-  } catch (error) {
-    // Handle errors that occur during logout
-    console.error('Logout failed', error);
-    throw new Error('Logout failed'); // Re-throw the error to fail the test
   }
 }
