+++ src/e2e/tests/utils/index.ts
@@ -1,0 +1,14 @@
+//Auth logins and logout
+export { cpoLogin, logout } from './auth/cpoAuth';
+export { mpoLogin } from './auth/mpoAuth';
+export { drafterLogin } from './auth/drafterAuth';
+
+//Env variables
+export {
+  drafterUsername,
+  drafterPassword,
+  mpoUsername,
+  mpoPassword,
+  cpoUsername,
+  cpoPassword,
+} from './env/env';
